+++ app/scripts/reducers/granules.js
@@ -24,6 +24,10 @@ import {
   GRANULE_REINGEST_INFLIGHT,
   GRANULE_REINGEST_ERROR,
 
+  GRANULE_APPLYWORKFLOW,
+  GRANULE_APPLYWORKFLOW_INFLIGHT,
+  GRANULE_APPLYWORKFLOW_ERROR,
+
   GRANULE_REMOVE,
   GRANULE_REMOVE_INFLIGHT,
   GRANULE_REMOVE_ERROR,
@@ -55,6 +59,7 @@ export const initialState = {
   reprocessed: {},
   removed: {},
   reingested: {},
+  executed: {},
   deleted: {},
   recent: {}
 };
@@ -127,6 +132,18 @@ export default function reducer (state = initialState, action) {
     case GRANULE_REINGEST_ERROR:
       set(state, ['reingested', id, 'status'], 'error');
       set(state, ['reingested', id, 'error'], action.error);
+      break;
+
+    case GRANULE_APPLYWORKFLOW:
+      set(state, ['executed', id, 'status'], 'success');
+      set(state, ['executed', id, 'error'], null);
+      break;
+    case GRANULE_APPLYWORKFLOW_INFLIGHT:
+      set(state, ['executed', id, 'status'], 'inflight');
+      break;
+    case GRANULE_APPLYWORKFLOW_ERROR:
+      set(state, ['executed', id, 'status'], 'error');
+      set(state, ['executed', id, 'error'], action.error);
       break;
 
     case GRANULE_REMOVE: